Fix docs: appropriated -> relevant (#89)
Appropriate is about suitablity of smth in some context. We can say: It is appropriate to wear a tie on a wedding. Or we can say It is not appropriate to swear in PR comments

Relevant means smth is connected to some context. Like, this screenshot is relevant to this PR or This PR is about docs, so the screenshot is not relevant here

I think relevant would be a better choice here